requirements.txt
Django==2.2.6
django-dynamic-fixture==2.0.0
djangorestframework==3.10.3
black